Get TrueType benches work again
